+++ graphene_django_crud/__init__.py
@@ -5,7 +5,7 @@ from .types import DjangoGrapheneCRUD
 from .types import apply_where as where_input_to_Q
 
 
-__version__ = "1.1.0"
+__version__ = "1.1.1"
 
 __all__ = (
     "__version__",